refactor: move pkg/botel/system → pkg/util/healthutil
The botel/ package no longer contains any OTel code, so the name
is misleading. Moving the gopsutil-based health collector to
pkg/util/healthutil/ aligns it with the *util naming convention
(authutil, serverutil, storageutil, etc.) and deletes the empty
pkg/botel/ directory entirely.

No logic changes — pure rename.
