+++ internal/server/server.go
@@ -16,8 +16,8 @@ import (
 	"github.com/autobutler-org/autobutler/internal/db"
 	"github.com/autobutler-org/autobutler/internal/server/middleware"
 	"github.com/autobutler-org/autobutler/pkg/backup"
-	"github.com/autobutler-org/autobutler/pkg/botel/system"
 	"github.com/autobutler-org/autobutler/pkg/util/authutil"
+	"github.com/autobutler-org/autobutler/pkg/util/healthutil"
 	"github.com/autobutler-org/autobutler/pkg/util/deputil"
 	"github.com/autobutler-org/autobutler/pkg/util/eventbus"
 	"github.com/autobutler-org/autobutler/pkg/util/favoritesutil"
@@ -216,7 +216,7 @@ type StartOptions struct {
 }
 
 func StartServer(deps deputil.Dependencies, opts StartOptions) error {
-	systemCollector, err := system.Register()
+	systemCollector, err := healthutil.Register()
 	if err != nil {
 		return fmt.Errorf("failed to initialize system collector: %w", err)
 	}
